+++ include/workflow/task.hpp
@@ -9,7 +9,7 @@ using task_id = size_t;
 struct task {
     task_id const id;
     double const workload;
-    size_t const memory_requirement;
+    double const memory_requirement;
 };
 
 } // namespace workflow
